requirements.txt
flask
google-generative-ai
python-dotenv
gunicorn
